refactor: cleanup ledger popups (#4426)
* refactor: remove old ledger popups

* refactor: rename ledger popups

* refactor: fixes and renames
